Safari-611.1.10/LayoutTests/media/track/regions-webvtt/text-track-cue-region-attribute-expected.txt

Tests the region attribute of a cue.


EXPECTED (!!window.VTTRegion == 'true') OK

** Test the setter and getter through the JS API **
EXPECTED (cue.region == 'null') OK

** Test parsing a region attribute of a cue **
EXPECTED (typeof cue.region == 'object') OK

** Test that region attribute is ignored if either line position or cue size are specified or writing direction is not horizontal **
EXPECTED (cue.region == 'null') OK
EXPECTED (cue.region == 'null') OK
EXPECTED (cue.region == 'null') OK

END OF TEST
